ajout change privilege

+++ Care4Care/C4CApplication/views/BranchListRedirectView.py
@@ -7,7 +7,7 @@ from django.core.exceptions import PermissionDenied
 from django.core.urlresolvers import reverse_lazy
 
 
-class DeleteMemberFromBranchRedirectView(RedirectView):
+class BranchListRedirectView(RedirectView):
 
     url = ""
     user = None
@@ -16,12 +16,20 @@ class DeleteMemberFromBranchRedirectView(RedirectView):
         if 'email' not in self.request.session:
             raise PermissionDenied  # HTTP 403
         self.user = create_user(self.request.session['email'])
-        return super(DeleteMemberFromBranchRedirectView, self).dispatch(request, *args, **kwargs)
+        return super(BranchListRedirectView, self).dispatch(request, *args, **kwargs)
 
     @never_cache
     def get(self, request, *args, **kwargs):
         mail_member = kwargs['mail']
         branch_name = kwargs['branch']
+        action = kwargs['action']
+        
         self.url = reverse_lazy("memberlist", kwargs={'pk': branch_name})
-        self.user.delete_member_from_branch(branch_name, mail_member)
-        return super(DeleteMemberFromBranchRedirectView, self).get(request, *args, **kwargs)
+        if action=='0' :  #Remove the member from the branch
+            self.user.delete_member_from_branch(branch_name, mail_member)
+        elif action=='1' :    #Promote/de-promote to verified
+            self.user.modify_tag_member(mail_member, 4)
+        else : 
+            self.user.modify_tag_member(mail_member, 8)
+            
+        return super(BranchListRedirectView, self).get(request, *args, **kwargs)
